Reshuffle unit tests
